requirements.txt
streamlit==1.28.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.7,<0.2
openai>=1.6.1
python-dotenv>=1.0.0
pypdf>=3.17.4
faiss-cpu>=1.7.4
tiktoken>=0.5.2

